задание готово - 13. Личный проект: разделяй и властвуй (Часть 1)
